add menhir to the dune overlays
The reason for this is that the menhir git repo doesnt check
in the manual, and so the duniverse clone heuristics do not
permit grabbing the source archive (which is significantly
different from the git remote).

Therefore as a temporary measure I am checking in the menhir
source archive and adding the 3 opam files to the branch so
that duniverse embedding of menhir works.