Add dune port of opam-file-format 2.1.0
opam-file-format has a depopts on dune which doesn't satisfy
opam-monorepo's build system detection logic. We should probably
consider a better way to detect the build system, maybe using the
`build` field from the opam file but that wouldn't have necessarily
worked here either. A specific field in the opam file with the list of
supported build systems would be great but in the meantime this port
does the job.