Use tagged commit in io-page `2.3.0+dune`
As #141 has uncovered, the dune port of io-page just uses a git repo
with no fixed commit at all, so is entirely non-reproducible. As such at
the moment, anyone locking this version will get at least 2.4.0 which
comes with breaking changes compared to 2.3.0.

As such, this makes sure that io-page 2.3.0+dune is actually io-page
2.3.0.

Closes #142.